custom-llm/requirements.txt
python-dotenv ~= 1.0
requests ~= 2.0
